+++ libs/targets/gesha-rust-types/src/lib.rs
@@ -2,7 +2,7 @@ mod data_type;
 pub use data_type::DataType;
 
 mod definition;
-pub use definition::{Definition, ModDef, NewTypeDef, PresetDef, StructDef};
+pub use definition::Definition;
 
 mod definitions;
 pub use definitions::Definitions;
@@ -43,14 +43,23 @@ pub use keywords::KEYWORDS;
 mod media_type_def;
 pub use media_type_def::MediaTypeDef;
 
+mod mod_def;
+pub use mod_def::ModDef;
+
 mod module_declarations;
 pub use module_declarations::ModuleDeclarations;
 
 mod module_name;
 pub use module_name::ModuleName;
 
+mod new_type_def;
+pub use new_type_def::NewTypeDef;
+
 mod non_doc_comments;
 pub use non_doc_comments::NonDocComments;
+
+mod preset_def;
+pub use preset_def::PresetDef;
 
 mod request_body_def;
 pub use request_body_def::{MediaTypeVariant, MediaTypeVariants, RequestBodyDef};
@@ -60,6 +69,9 @@ pub use serde_attribute::SerdeAttribute;
 
 mod source_code;
 pub use source_code::SourceCode;
+
+mod struct_def;
+pub use struct_def::StructDef;
 
 mod struct_field;
 pub use struct_field::{StructField, StructFieldAttribute};
