+++ libs/targets/gesha-rust-types/src/mod_def.rs
@@ -1,0 +1,14 @@
+use crate::{Definition, Definitions, Imports, ModuleName};
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct ModDef {
+    pub name: ModuleName,
+    pub imports: Imports,
+    pub defs: Definitions,
+}
+
+impl From<ModDef> for Definition {
+    fn from(this: ModDef) -> Self {
+        Self::ModDef(this)
+    }
+}
