+++ libs/targets/gesha-rust-types/src/new_type_def.rs
@@ -1,0 +1,24 @@
+use crate::{DataType, Definition, TypeHeader};
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct NewTypeDef {
+    pub header: TypeHeader,
+    pub data_type: DataType,
+    _hide_default_constructor: bool,
+}
+
+impl NewTypeDef {
+    pub fn new(header: TypeHeader, data_type: DataType) -> Self {
+        Self {
+            header,
+            data_type,
+            _hide_default_constructor: true,
+        }
+    }
+}
+
+impl From<NewTypeDef> for Definition {
+    fn from(x: NewTypeDef) -> Self {
+        Self::NewTypeDef(x)
+    }
+}
